Terminal client support for remote libraries

+++ src/plugins/httpdatastream/LruDiskCache.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////////
 //
-// Copyright (c) 2004-2019 musikcube team
+// Copyright (c) 2004-2020 musikcube team
 //
 // All rights reserved.
 //
@@ -36,6 +36,7 @@
 #include <vector>
 #include <mutex>
 #include <ctime>
+#include <cstdint>
 
 #include <boost/filesystem.hpp>
 
@@ -56,7 +57,7 @@ class LruDiskCache {
 
     private:
         struct Entry {
-            size_t id;
+            uint64_t id;
             std::string path;
             std::string type;
             std::time_t time;
